[ALPS02542738] Audio: Olympus Driver Porting for 32bit
[Detail]
    Use new architecture on Olympus for 32bit.
[Solution]
    1. Set new architecture config.
    2. Update config control table.
    3. resolve build break issues for 32bit platform

MTK-Commit-Id: 003e7ed0679ee8bdab1de2903e9f4ca5ce69a270

Feature: Audio Playback(*)
CR-Id: ALPS02542738

+++ sound/soc/mediatek/common/mt_soc_pcm_anc.c
@@ -776,7 +776,7 @@ static ssize_t ANCService_write(struct file *fp, const char __user *data,
 {
 	char temp_str[MAX_TABS * 4];
 
-	pr_debug("ANCService_write write count %d", (int)count);
+	pr_debug("ANCService_write write count %zu", count);
 
 	if (count > MAX_TABS * 4)
 		return -EINVAL;
